+++ src/undoers/add_palette.h
@@ -32,7 +32,7 @@ public:
   AddPalette(undo::ObjectsContainer* objects, Sprite* sprite, int palette_frame);
 
   void dispose() OVERRIDE;
-  int getMemSize() const OVERRIDE { return sizeof(*this); }
+  size_t getMemSize() const OVERRIDE { return sizeof(*this); }
   void revert(undo::ObjectsContainer* objects, undo::UndoersCollector* redoers) OVERRIDE;
 
 private:
