+++ src/undoers/dirty_area.h
@@ -35,7 +35,7 @@ public:
   DirtyArea(undo::ObjectsContainer* objects, Image* image, Dirty* dirty);
 
   void dispose() OVERRIDE;
-  int getMemSize() const OVERRIDE { return sizeof(*this) + getStreamSize(); }
+  size_t getMemSize() const OVERRIDE { return sizeof(*this) + getStreamSize(); }
   void revert(undo::ObjectsContainer* objects, undo::UndoersCollector* redoers) OVERRIDE;
 
 private:
