+++ src/undoers/remove_image.h
@@ -34,7 +34,7 @@ public:
   RemoveImage(undo::ObjectsContainer* objects, Stock* stock, int imageIndex);
 
   void dispose() OVERRIDE;
-  int getMemSize() const OVERRIDE { return sizeof(*this) + getStreamSize(); }
+  size_t getMemSize() const OVERRIDE { return sizeof(*this) + getStreamSize(); }
   void revert(undo::ObjectsContainer* objects, undo::UndoersCollector* redoers) OVERRIDE;
 
 private:
